Refactor submitter CLI and migrate database migrations to lib directory
