+++ packages/submitter/lib/cli/commands/generateDatabaseTypes.ts
@@ -1,0 +1,32 @@
+import { join } from "node:path"
+import { KyselyBunSqliteDialect, generate } from "kysely-codegen"
+import { db } from "#lib/database"
+import { overrides } from "#lib/database/type-generation-overrides"
+
+/**
+ * This generates all the typescript types definitions based on the current database state, and save
+ * it to `lib/database/generated.d.ts`. It should be run after migrations are applied.
+ *
+ * Note: This file can't easily be built for production, but the type generation is also
+ * not needed in production. It
+ */
+
+const baseDir = join(import.meta.dir, "../../../")
+
+const typeGenOutFile = "lib/database/generated.d.ts"
+
+const typeCodeGenOptions = {
+    camelCase: false,
+    db,
+    dialect: new KyselyBunSqliteDialect(),
+    // logger: new Logger(),
+    outFile: join(baseDir, typeGenOutFile),
+    overrides,
+    singular: true,
+    verify: false,
+}
+
+export async function generateDatabaseTypesCommand() {
+    await generate(typeCodeGenOptions)
+    console.log(`\nSuccessfully Generated types: ${typeGenOutFile}`)
+}
