+++ packages/submitter/lib/cli/index.ts
@@ -1,19 +1,41 @@
-import { migrateCommand } from "./commands/migrate"
-import { routesCommand } from "./commands/routes"
 import { getCommand, showHelp } from "./utils"
 
 export async function processCmd(argv: string[]) {
-    const { values, positionals, command } = getCommand(argv, ["routes", "migrate"] as const)
+    const { values, positionals, command } = getCommand(argv, ["routes", "migrate", "db-types"] as const)
+    if (!command.success) {
+        if (command.data) console.error(`Unknown command: ${command.data}`)
+        if (command.data && command.error) console.error(`  => ${command.error?.flatten().formErrors[0]}`)
+        showHelp()
+        process.exit(1)
+    }
+
     switch (command.data) {
-        case "routes":
-            routesCommand({ values, positionals, command })
+        case "routes": {
+            try {
+                const { routesCommand } = await import("./commands/routes")
+                routesCommand({ values, positionals, command })
+            } catch {
+                console.error("Error loading routes command. Ensure 'zod-openapi' is installed.")
+                process.exit(1)
+            }
             break
-        case "migrate":
+        }
+        case "migrate": {
+            const { migrateCommand } = await import("./commands/migrate")
             migrateCommand({ values, positionals, command })
             break
+        }
+        case "db-types": {
+            try {
+                const { generateDatabaseTypesCommand } = await import("./commands/generateDatabaseTypes")
+                generateDatabaseTypesCommand()
+            } catch {
+                console.error("Error generating typescript definitions. Ensure 'kysely-codegen' is installed.")
+                process.exit(1)
+            }
+            break
+        }
         default: {
-            if (command.data) console.error(`Unknown command: ${command.data}`)
-            if (command.data && command.error) console.error(`  => ${command.error?.flatten().formErrors[0]}`)
             showHelp()
             process.exit(1)
         }
